moved config file
